requirements.txt
django
pillow
django-active-link
django-filter